Add input class

+++ StreamlineCore/src/IO/KeyCodes.h
@@ -4,7 +4,7 @@
 
 namespace slc {
 
-	using KeyCode = uint16_t;
+	using KeyCode = int;
 
 	namespace Key
 	{
