+++ StreamlineCore/src/IO/MouseCodes.h
@@ -4,7 +4,7 @@
 
 namespace slc {
 
-	using MouseCode = uint16_t;
+	using MouseCode = int;
 
 	namespace Mouse
 	{
